+++ crates/runner-host/src/board/platform.rs
@@ -28,6 +28,8 @@ impl Api for Impl {
     #[cfg(feature = "usb")]
     type Protocol = crate::board::usb::ProtocolImpl;
 
+    type Update = UpdateImpl;
+
     fn serial() -> Cow<'static, [u8]> {
         from_hex(option_env!("WASEFIRE_HOST_SERIAL"))
     }
@@ -37,6 +39,25 @@ impl Api for Impl {
     }
 
     fn reboot() -> Result<!, Error> {
+        Err(Error::world(Code::NotImplemented))
+    }
+}
+
+pub enum UpdateImpl {}
+impl wasefire_board_api::Support<bool> for UpdateImpl {
+    const SUPPORT: bool = false;
+}
+impl wasefire_board_api::platform::update::Api for UpdateImpl {
+    fn metadata() -> Result<Box<[u8]>, Error> {
+        Err(Error::world(Code::NotImplemented))
+    }
+    fn initialize(_dry_run: bool) -> Result<(), Error> {
+        Err(Error::world(Code::NotImplemented))
+    }
+    fn process(_chunk: &[u8]) -> Result<(), Error> {
+        Err(Error::world(Code::NotImplemented))
+    }
+    fn finalize() -> Result<(), Error> {
         Err(Error::world(Code::NotImplemented))
     }
 }