+++ crates/scheduler/src/event/platform.rs
@@ -15,13 +15,11 @@
 use wasefire_board_api::platform::Event;
 use wasefire_board_api::Api as Board;
 
-#[cfg(feature = "board-api-platform-protocol")]
 pub mod protocol;
 
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 #[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
 pub enum Key {
-    #[cfg(feature = "board-api-platform-protocol")]
     Protocol(protocol::Key),
 }
 
@@ -34,7 +32,6 @@ impl<B: Board> From<Key> for crate::event::Key<B> {
 impl<'a> From<&'a Event> for Key {
     fn from(event: &'a Event) -> Self {
         match *event {
-            #[cfg(feature = "board-api-platform-protocol")]
             Event::Protocol(ref event) => Key::Protocol(event.into()),
         }
     }
@@ -42,7 +39,6 @@ impl<'a> From<&'a Event> for Key {
 
 pub fn process(event: Event) {
     match event {
-        #[cfg(feature = "board-api-platform-protocol")]
         Event::Protocol(_) => protocol::process(),
     }
 }